backend/requirements.txt
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
qrcode[pil]>=7.4.2
ultralytics>=8.0.0
pyttsx3>=2.90
numpy>=1.24.0
